feat: main 페이지 info 수정
