+++ client/src/Components/atom/Calendar/index.ts
@@ -1,11 +1,11 @@
 import Component, { StateType } from '@/core/Component';
 import jsx from '@/core/jsx';
-import { AllHistorytype, HistoriesType } from '@/Pages/Main';
 import { getState, subscribe } from '@/core/observer';
 import { userState } from '@/Model';
 import dayjs from 'dayjs';
 import './style';
 import { returnPrice } from '@/utils/util';
+import { AllHistorytype, HistoriesType } from '@/shared/type';
 
 interface CalendarProps {
   firstDay: Date;
